fix: silentLoginIframe Timeouts.
signinCallback wasn't being called when state was set but not code,
which happens when there are errors. I also did some error handling
the refreshToken calll. This is kind of a quick fix. I need to revisit the
flows here to make sure all the correct things are being done for the
various callback states.